==> 1_Energy_data_exploration_and_cleaning.ipynb <==
Here we will explore the data, get to know the data types, and if there are any inconsistencies, based from the data documentation given.

import pandas as pd

url = "Your_Google_drive_link" ## add your own google drive infor
path = "https://drive.google.com/uc?export=download&id="+url.split("/")[-2]
MES_data = pd.read_csv(path)

MES_data_1 = MES_data.copy()
MES_data_1 = pd.DataFrame(MES_data)

MES_data_1.info()

# drop information not needed for analysis and renamed columns

MES_data_1.head(10)

MES_data_1 = MES_data_1.drop([0,1,2,3,4,5,6,7,8],axis = 0).copy()

MES_data_1 = MES_data_1.rename( columns = {'The Monthly electricity data explorer and the main highlights are available here: https://www.iea.org/reports/monthly-electricity-statistics-overview':'country',
                              'Unnamed: 1':'date','Unnamed: 2':'balance', 'Unnamed: 3':'product', 'Unnamed: 4':'value', 'Unnamed: 5':'unit'}).copy()






# check column types, for duplicates, null, values,
MES_data_1.info() # need to transform data to year and month, and value to int64.

MES_data_1.duplicated() # zero duplicate enteries
MES_data_1.describe()

MES_data_1.isna().sum() # 51 null values -
MES_data_1['value'].isna().value_counts(normalize=True)


# drop na values
MES_data_1 = MES_data_1.dropna(axis=0).copy()
MES_data_1.isna().sum()


MES_data_1.head(50)



Need to think about if we want to get the data only for one country and if we shold split this up into different tables to pull information from.

Also want to konw all the other catagories for product.



# Need to change date to an actual data and value to int64
MES_data_1['date']= pd.to_datetime(MES_data_1['date'].astype(str), format='%B %Y')
MES_data_1['date']



MES_data_1['value'] = pd.to_numeric(MES_data_1['value'], downcast='float')
MES_data_1['value']

MES_data_1.to_csv("MES_data_1_clean.csv", index=False)
