==> 2_Get_city_coordinates.ipynb <==
## Getting Citiy cooridinates

In order to get city information and population data via API NINJA we need the coordinates to so we are going throught that process here. We had trouble running the extraction in Juptyer lab so we switched to google collab.



import requests
import pandas as pd

API_key = "BEaP002yLXL0xoSywBLwWg==PJDPv4SmopSGvbEQ"

# it was decided later in the process that using 1 city climate data was easier in facilitating modeling and analysis.
cities = [
    'Berlin', 'Dortmund', 'Kassel', 'Munich',
    'Madrid', 'Bilbao', 'Seville', 'Valencia',
    'Oslo', 'Trondheim', 'Harstad', 'Alta',
    'Bratislava', 'Banksa', 'Kosice', 'Zilina'
]

countries = [
    'DE', 'DE', 'DE', 'DE',
    'ES', 'ES', 'ES', 'ES',
    'NO', 'NO', 'NO', 'NO',
    'SK', 'SK', 'SK', 'SK'
]

url = "https://api.api-ninjas.com/v1/geocoding"

cities_list = []
lat_list = []
lon_list = []
country_code_list = []

for c, country in zip(cities, countries):
    response = requests.get(
        url,
        params={'city': c, 'country': country},
        headers={'X-Api-Key': API_key}
    )
    data = response.json()

    if data:
        item = data[0]   # first match

        cities_list.append(item['name'])
        lat_list.append(item['latitude'])
        lon_list.append(item['longitude'])
        country_code_list.append(item['country'])


print(cities_list)
print(lat_list)
print(lon_list)
print(country_code_list)


city_dict = {
    c: {"lat": lat, "lon": lon, "country": country}
    for c, lat, lon, country in zip(cities_list, lat_list, lon_list, country_code_list)
}



city_geo = pd.DataFrame(city_dict)
city_geo = city_geo.transpose()
city_geo

# city_geo[city] = city_geo[]
city_geo = city_geo.reset_index()
# city_geo
#city_geo = city_geo.drop(columns = 'level_0').copy()
city_geo = pd.DataFrame.rename(city_geo, columns = {'index':'city','lat	':'lat', 'lon':'lon','country':'country_code'})
city_geo

city_geo.to_csv('city_coordinates.csv')

Run data in collab as it has a better serve
