==> 5_Weather_data_cleaning_preperation.ipynb <==
import pandas as pd
import requests

!pip install sqlalchemy
!pip install pymysql

schema = "EuroEnergy"
host = "127.0.0.1"
user = "root"
password = "YOUR PASSWORD"
port = 3306

connection_string = f'mysql+pymysql://{user}:{password}@{host}:{port}/{schema}'

 weather_data = pd.read_csv('all_cities_df.csv')
 weather_data.head()

city_coordinates = pd.read_csv('city_coordinates.csv')
city_coordinates = city_coordinates.drop(columns = 'Unnamed: 0').copy()

weather_data.info()
city_coordinates

country_info = pd.read_sql('countries', con=connection_string)

country_info



#country_map = {'DE':'Germany', 'ES': 'Spain', 'NO':'Norway', 'SK': 'Slovak Republic'
city_coordinates['country'] = ''

city_coordinates['country'] = ''  # Initialize empty string column

for idx, row in city_coordinates.iterrows():
    code = row['country_code']
    if code == 'DE':
        city_coordinates.at[idx, 'country'] = 'Germany'
    elif code == 'ES':
        city_coordinates.at[idx, 'country'] = 'Spain'
    elif code == 'NO':
        city_coordinates.at[idx, 'country'] = 'Norway'
    elif code == 'SK':
        city_coordinates.at[idx, 'country'] = 'Slovak Republic'

country_cities = city_coordinates.merge(country_info, left_on= 'country',
                                        right_on='country_name',
                                        how = 'left').drop(axis = 0, columns ='country')
country_cities

# MySQL code

USE EuroEnergy;

CREATE TABLE country_cities (
	city_id INT AUTO_INCREMENT,
    city VARCHAR (250) NOT NULL,
    lat FLOAT NOT NULL,
    lon FLOAT NOT NULL,
    country_code VARCHAR(3)NOT NULL,
    country_id INT NOT NULL,
    country_name VARCHAR(250) NOT NULL,
    PRIMARY KEY (city_id),
    FOREIGN KEY (country_id) References countries(country_id)
    );
    

country_cities.to_sql('country_cities',
                  if_exists='append',
                  con=connection_string,
                  index=False)

country_cities = pd.read_sql('country_cities', con = connection_string)

country_cities


 country_weather_data = weather_data.merge(country_cities, left_on = ['lat','lon'],
                    right_on=['lat','lon'],
                    how = 'left').drop(axis = 0, columns = 'country_code')
 country_weather_data = country_weather_data.drop( axis= 0, columns = ['city_x']).copy()
 country_weather_data = country_weather_data.rename( columns = {'datetime':'datetime',
                'tempmax':'tempmax','tempmin':'tempmin',
                'temp':'temp','humidity':'humidity', 'precip':'precip', 'windspeed':'windspeed',
            'cloudcover':'cloudcover', 'conditions':'conditions', 'lat':'lat',
        'lon':'lon', 'city_y':'city', 'country_id':'country_id',
        'country_name':'country_name'})

country_weather_data

country_weather_data.pd.to_csv('daily_weather_data'.csv)

#country_weather_data.to_csv('daily_weather_data.csv')
country_weather_data = pd.read_csv('daily_weather_data.csv')

weather_data['lat'] = weather_data['lat'].round(2)
weather_data['lon'] = weather_data['lon'].round(2)

country_cities['lat'] = country_cities['lat'].round(2)
country_cities['lon'] = country_cities['lon'].round(2)

country_cities

 country_weather_data.isna().sum()


#country_weather_data['city'].unique()

country_weather_data['datetime'] = pd.to_datetime(country_weather_data['datetime'])
country_weather_data['year_month'] = country_weather_data['datetime'].dt.to_period('M')

country_weather_data['year_month'].unique()


country_weather_data['year_month'] = country_weather_data['datetime'].dt.to_period('M')

country_weather_data['year_month'].unique()

#monthly weather aggergates
monthly_stats = (
    country_weather_data.groupby(['year_month', 'city','country_name'])
      .agg(
          avg_temp=('temp', 'mean'),
          # most_temp = ('temp', 'mode'),
          min_temp=('tempmin', 'min'),
                  max_temp=('tempmax', 'max'),
                  max_windspeed = ('windspeed', 'max'),
                  avg_windspeed = ('windspeed', 'mean'),
                  avg_cloudcover = ('cloudcover', 'mean'),

                  total_precip=('precip', 'sum'),
                  #dominant_temp_range=('temp_range', lambda x: x.mode()[0])
      )
      .reset_index()
)


Country_monthly_weather.to_csv('monthly_weather_stats.csv')

#Country_monthly_weather = pd.read_csv('monthly_weather_stats.csv')
Country_monthly_weather = Country_monthly_weather.drop(axis = 0, columns = 'Unnamed: 0').copy()
# monthly_stats
# country_cities
Country_monthly_weather

Country_monthly_weather = monthly_stats.merge(country_cities, on = ['city','country_name'],
                    how = 'left').drop(axis = 0, columns = ['lat','lon']).copy()

Next step, merge with country_table info to get country_id,
Send to MySQL
Set-energy data as period
Merge with Energy data using period, and country


Energy_data = pd.read_csv('energy_population_data.csv')
Energy_data=Energy_data.drop(axis = 0, columns = 'Unnamed: 0').copy()
Energy_data

Energy_data['date_of_energy_records'] = pd.to_datetime(Energy_data['date_of_energy_records'])
Energy_data['year_month'] = Energy_data['date_of_energy_records'].dt.to_period('M')

Energy_data['year_month'].unique()
# Country_monthly_weather['year_month']pd.to

Country_monthly_weather['year_month'] = (
    Country_monthly_weather['year_month'].astype('period[M]')
)

Energy_data['year_month'] = (
    Energy_data['year_month'].astype('period[M]')
)

energy_weather_population = Country_monthly_weather.merge(Energy_data,
                              left_on = ['country_name', 'year_month'],
                              right_on = ['country', 'year_month'],
                              how = 'left').drop (axis = 0 ,
                                                  columns = ['country', 'year','Pecent_population_urban','country_code'])

energy_weather_population.info()
energy_weather_population.to_csv('energy_weather_population.csv')
